Added preprocessing script with torchrun

+++ tools/preprocess_data.py
@@ -1,30 +1,14 @@
 import argparse
-import multiprocessing
 import os
 import shutil
+import sys
 
 import numpy as np
+import torch.distributed as dist
 from tqdm import tqdm
-from transformers import AutoTokenizer, PreTrainedTokenizerBase
+from transformers import AutoTokenizer
 
-from datasets import Dataset, concatenate_datasets, load_dataset
-
-
-def preprocess_shard(
-    dataset_shard: Dataset, output_file: str, tokenizer: PreTrainedTokenizerBase, column: str, add_special_tokens: bool
-):
-    dataset_shard = dataset_shard.map(
-        lambda x: {"input_ids": tokenizer(x, add_special_tokens=add_special_tokens).input_ids},
-        input_columns=column,
-        batched=True,
-        desc="Tokenizing Dataset",
-        remove_columns=[column],
-    )
-    input_ids_file = open(output_file, "wb")
-    for sample in dataset_shard:
-        np_array = np.array(sample["input_ids"], dtype=np.uint16)
-        input_ids_file.write(np_array.tobytes(order="C"))
-    input_ids_file.close()
+from datasets import concatenate_datasets, load_dataset
 
 
 def get_args():
@@ -52,15 +36,18 @@ def get_args():
     group = parser.add_argument_group(title="output data")
     group.add_argument("--output-prefix", type=str, required=True, help="Path to the output processed dataset file")
 
-    group = parser.add_argument_group(title="runtime")
-    group.add_argument("--num-workers", type=int, default=8, help="Number of workers processing the dataset")
-
     args = parser.parse_args()
 
     return args
 
 
 def main(args):
+
+    world_size, rank = int(os.environ["WORLD_SIZE"]), int(os.environ["RANK"])
+
+    # Remove stdout from all processes except main to not flood the stdout
+    if rank:
+        sys.stdout = open(os.devnull, "w")
 
     # Check if output directory exists
     if not os.path.isdir(os.path.abspath(os.path.join(args.output_prefix, os.path.pardir))):
@@ -73,8 +60,9 @@ def main(args):
             [ds[splits] for splits in ds.keys()]
         )  # load_dataset returns DatasetDict and we want a Dataset
     else:
-        ds = load_dataset(args.input, split=args.split, num_proc=args.num_workers)
+        ds = load_dataset(args.input, split=args.split)
 
+    ds = ds.shard(num_shards=world_size, index=rank, contiguous=True)
     ds = ds.select_columns(args.column)
 
     tokenizer = AutoTokenizer.from_pretrained(args.tokenizer_name_or_path)
@@ -82,43 +70,45 @@ def main(args):
     # Create tmp directory for worker outputs
     tmp_folder = os.path.abspath(os.path.join(args.output_prefix, os.pardir, "tmp"))
     os.makedirs(tmp_folder, exist_ok=True)
-    workers_output_files = []
-    processes = []
 
-    print("Creating worker output files...")
-    for worker in range(args.num_workers):
-        worker_output_file = os.path.join(tmp_folder, f"worker_{worker}_input_ids.npy")
-        workers_output_files.append(worker_output_file)
+    print("Creating workers output files...")
+    worker_output_file = os.path.join(tmp_folder, f"worker_{rank}_input_ids.npy")
+    ds = ds.map(
+        lambda x: {"input_ids": tokenizer(x, add_special_tokens=args.add_special_tokens).input_ids},
+        input_columns=args.column,
+        batched=True,
+        desc="Tokenizing Dataset",
+        remove_columns=[args.column],
+    )
 
-        p = multiprocessing.Process(
-            target=preprocess_shard,
-            args=(
-                ds.shard(num_shards=args.num_workers, index=worker, contiguous=True),
-                worker_output_file,
-                tokenizer,
-                args.column,
-                args.add_special_tokens,
-            ),
-        )
+    worker_input_ids_file = open(worker_output_file, "wb")
+    for sample in ds:
+        np_array = np.array(sample["input_ids"], dtype=np.uint16)
+        worker_input_ids_file.write(np_array.tobytes(order="C"))
+    worker_input_ids_file.close()
 
-        p.start()
-        processes.append(p)
+    # Wait for all workers to process each shard of the Dataset
+    dist.barrier()
 
-    for p in processes:
-        p.join()
+    # Only the main rank merges the worker files
+    if not rank:
+        output_file = f"{args.output_prefix}_input_ids.npy"
+        input_ids_file = open(output_file, "wb")
+        for worker_idx in tqdm(range(world_size), desc="Merging workers output files"):
+            worker_output_file = os.path.join(tmp_folder, f"worker_{worker_idx}_input_ids.npy")
+            with open(worker_output_file, "rb") as f:
+                shutil.copyfileobj(f, input_ids_file)
+            os.remove(worker_output_file)
 
-    output_file = f"{args.output_prefix}_input_ids.npy"
-    input_ids_file = open(output_file, "wb")
-    for worker_output_file in tqdm(workers_output_files, desc="Merging worker output files"):
-        with open(worker_output_file, "rb") as f:
-            shutil.copyfileobj(f, input_ids_file)
-        os.remove(worker_output_file)
+        input_ids_file.close()
+        os.rmdir(tmp_folder)
+        print(f"Done! {args.input} processed dataset stored in {output_file}")
 
-    input_ids_file.close()
-    os.rmdir(tmp_folder)
-    print(f"Done! {args.input} processed dataset stored in {output_file}")
+    else:  # Close devnull stdout redirect
+        sys.stdout.close()
 
 
 if __name__ == "__main__":
     _args = get_args()
+    dist.init_process_group(backend="gloo")
     main(_args)
